API remove Object::useCustomClass API Broke out Injectable, Configurable, and Extensible from Object into traits BUG Fix missing extraDataObjects in tests
