+++ docs/en/04_Changelogs/4.0.0.md
@@ -23,6 +23,7 @@
  * `Upload::load` now stores assets directly without saving into a `File` dataobject.
  * Protected file storage is now a core Framework API. See [/developer_guides/files/file_security] for
    more information.
+ * `Object::useCustomClass` has been removed. You should use the config API with Injector instead.
 
 ## New API
 
@@ -38,6 +39,10 @@
  * `AssetControlExtension` is applied by default to all DataObjects, in order to support the management
    of linked assets and file protection.
  * `ProtectedFileController` class is used to serve up protected assets.
+ * `Object` has been broken up into various traits, each of which can be added to other objects independently:
+    * `Configurable` Provides Config API helper methods
+    * `Injectable` Provides Injector API helper methods
+    * `Extensible` Allows extensions to be applied
 
 ### Front-end build tooling for CMS interface
 
@@ -62,6 +67,10 @@ or developing your own website. These improvements are mainly geared at CMS core
  * Added SilverStripeComponent base class for building React UIs
 
 ## Deprecated classes/methods
+
+### Core
+
+ * `debugmethods` querystring argument has been removed from debugging.
 
 ### ORM
